RxProps: Separates live resolver logic into "makeReactive" util
